Move substitutions tests to separate file
