+++ src/tools/illink/test/Mono.Linker.Tests.Cases/Substitutions/Dependencies/TestFeatures.cs
@@ -1,0 +1,12 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+namespace ILLink.RoslynAnalyzer
+{
+	public class TestFeatures
+	{
+		public static bool IsUnreferencedCodeSupported => true;
+
+		public static bool IsAssemblyFilesSupported => true;
+	}
+}
