+++ src/tools/illink/test/Mono.Linker.Tests.Cases/Substitutions/FeatureGuardSubstitutions.xml
@@ -1,7 +1,7 @@
 <linker>
     <assembly fullname="test, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null">
         <!-- Override the substitutions implied by FeatureGuard/FeatureSwitchDefinition. -->
-        <type fullname="Mono.Linker.Tests.Cases.DataFlow.FeatureGuardAttributeDataFlow/FeatureGuardPrecedence">
+        <type fullname="Mono.Linker.Tests.Cases.Substitutions.FeatureGuardSubstitutions/FeatureGuardPrecedence">
             <method signature="System.Boolean get_GuardWithXml()" body="stub" value="true" />
             <method signature="System.Boolean get_SwitchWithXml()" body="stub" value="true" />
             <method signature="System.Boolean get_GuardAndSwitchWithXml()" body="stub" value="true" />
